performance fixes/improvements, bugfix

+++ steghide/src/BmpSampleValue.cc
@@ -61,10 +61,10 @@ float BmpSampleValue::calcDistance (const SampleValue *s) const
 	But calcDistance is called very often, a dynamic_cast costs a lot of time and
 	it does not make sense to pass anything but a BmpSampleValue as s anyway. */
 
-	float dr = (float) getRed() - (float) sample->getRed() ;
-	float dg = (float) getGreen() - (float) sample->getGreen() ;
-	float db = (float) getBlue() - (float) sample->getBlue() ;
-	return (sqrt (dr*dr + dg*dg + db*db)) ;
+	int dr = (int) getRed() - (int) sample->getRed() ;
+	int dg = (int) getGreen() - (int) sample->getGreen() ;
+	int db = (int) getBlue() - (int) sample->getBlue() ;
+	return (sqrt ((float) (dr*dr + dg*dg + db*db))) ;
 }
 
 float BmpSampleValue::getRadius () const
